+++ tests/testorc.py
@@ -26,8 +26,11 @@ import datetime
 from datetime import timezone
 import os
 
+import numpy as np
+import pyarrow
+import pyarrow.orc
+
 import scrapy
-import pyorc
 from scrapy.loader import ItemLoader
 from zuinnote.scrapy.contrib.bigexporters import OrcItemExporter
 from .testitem import TestItem
@@ -56,20 +59,42 @@ class TestOrcItemExporter(unittest.TestCase):
         # create exporter
         itemExporter = OrcItemExporter(
             file=self.file,
-            compression=pyorc.CompressionKind.ZLIB,
-            compressionstrategy=pyorc.CompressionStrategy.SPEED,
-            blocksize=65536,
-            batchsize=1024,
-            stripesize=67108864,
-            recordcache=10000,
-            schemastring="struct<ftext:string,ftext_array:array<string>,ffloat:float,fint:int,fbool:boolean,fdatetime:timestamp>",
+            no_items_batch=10000,
             convertallstrings=False,
-            bloomfiltercolumns=None,
-            bloomfilterfpp=0.05,
-            converters=None,
-            metadata=None,
+            batch_size=1024,
+            stripe_size=64 * 1024 * 1024,
+            compression="zstd",
+            compression_block_size=64 * 1024,
+            compression_strategy="speed",
+            row_index_stridex=10000,
+            padding_tolerance=0.0,
+            dictionary_key_size_threshold=0.0,
+            bloom_filter_columns=None,
+            bloom_filter_fpp=0.05,
         )
         self.export_type_schema(itemExporter)
+
+    def test_orc_export_string_schema(self):
+        """
+        Test if file is correctly written with strings
+        """
+        # create exporter
+        itemExporter = OrcItemExporter(
+            file=self.file,
+            no_items_batch=10000,
+            convertallstrings=True,
+            batch_size=1024,
+            stripe_size=64 * 1024 * 1024,
+            compression="zstd",
+            compression_block_size=64 * 1024,
+            compression_strategy="speed",
+            row_index_stridex=10000,
+            padding_tolerance=0.0,
+            dictionary_key_size_threshold=0.0,
+            bloom_filter_columns=None,
+            bloom_filter_fpp=0.05,
+        )
+        self.export_string_schema(itemExporter)
 
     def test_orc_export_type_schema_recordcache(self):
         """
@@ -78,18 +103,18 @@ class TestOrcItemExporter(unittest.TestCase):
         # create exporter
         itemExporter = OrcItemExporter(
             file=self.file,
-            compression=pyorc.CompressionKind.ZLIB,
-            compressionstrategy=pyorc.CompressionStrategy.SPEED,
-            blocksize=65536,
-            batchsize=1024,
-            stripesize=67108864,
-            recordcache=3,
-            schemastring="struct<ftext:string,ftext_array:array<string>,ffloat:float,fint:int,fbool:boolean,fdatetime:timestamp>",
+            no_items_batch=3,
             convertallstrings=False,
-            bloomfiltercolumns=None,
-            bloomfilterfpp=0.05,
-            converters=None,
-            metadata=None,
+            batch_size=1024,
+            stripe_size=64 * 1024 * 1024,
+            compression="zstd",
+            compression_block_size=64 * 1024,
+            compression_strategy="speed",
+            row_index_stridex=10000,
+            padding_tolerance=0.0,
+            dictionary_key_size_threshold=0.0,
+            bloom_filter_columns=None,
+            bloom_filter_fpp=0.05,
         )
         self.export_type_schema(itemExporter)
 
@@ -113,41 +138,99 @@ class TestOrcItemExporter(unittest.TestCase):
         itemExporter.finish_exporting()
         self.file.close()
         # reread file and compare results
-        with open(self.filename, "rb") as f:
-            reader = pyorc.Reader(f, struct_repr=pyorc.StructRepr.DICT)
+        rcount = 0
+        df = pyarrow.orc.read_table(self.filename).to_pandas()
+        for indx, record in df.iterrows():
             self.assertEqual(
-                len(reader),
-                10,
-                msg="Number of records read corresponds to number of records written",
+                "this is a test text",
+                record.get("ftext", None),
+                msg="String data is read correctly",
             )
-            for i in range(len(reader)):
-                record = next(reader)
-                self.assertEqual(
-                    "this is a test text",
-                    record.get("ftext", None),
-                    msg="String data is read correctly",
-                )
-                self.assertEqual(
+            self.assertIsNone(
+                np.testing.assert_array_equal(
                     ["test1", "test2", "test3", "test4"],
                     record.get("ftext_array", None),
-                    msg="String array data is read correctly",
-                )
-                self.assertEqual(
-                    float(2.5),
-                    record.get("ffloat", None),
-                    msg="Float data is read correctly",
-                )
-                self.assertEqual(
-                    int(10), record.get("fint", None), msg="Int data is read correctly"
-                )
-                self.assertFalse(
-                    record.get("fbool", None), msg="Bool data is read correctly"
-                )
-                self.assertEqual(
-                    datetime_obj,
-                    record.get("fdatetime", None),
-                    msg="DateTime data is read correctly",
-                )
+                ),
+                msg="String array data is read correctly",
+            )
+            self.assertEqual(
+                float(2.5),
+                record.get("ffloat", None),
+                msg="Float data is read correctly",
+            )
+            self.assertEqual(
+                int(10), record.get("fint", None), msg="Int data is read correctly"
+            )
+            self.assertFalse(
+                record.get("fbool", None), msg="Bool data is read correctly"
+            )
+            self.assertEqual(
+                datetime_str,
+                record.get("fdatetime", None).strftime(datetime_fmt),
+                msg="DateTime data is read correctly",
+            )
+            rcount += 1
+        self.assertEqual(
+            rcount,
+            10,
+            msg="Number of records read corresponds to number of records written",
+        )
+
+    def export_string_schema(self, itemExporter):
+        itemExporter.start_exporting()
+        # create and write some test data
+        num_records = 10
+        for i in range(num_records):
+            l = ItemLoader(TestItem())
+            l.add_value("ftext", "this is a test text")
+            l.add_value("ftext_array", ["test1", "test2", "test3", "test4"])
+            l.add_value("ffloat", float(2.5))
+            l.add_value("fint", int(10))
+            l.add_value("fbool", False)
+            datetime_str = "2020-02-29T11:12:13"
+            datetime_fmt = "%Y-%m-%dT%H:%M:%S"
+            datetime_obj = datetime.datetime.strptime(datetime_str, datetime_fmt)
+            l.add_value(
+                "fdatetime",
+                datetime_obj.replace(tzinfo=datetime.timezone.utc).timestamp(),
+            )
+            itemExporter.export_item(l.load_item())
+        itemExporter.finish_exporting()
+        self.file.close()
+        # reread file and compare results
+        rcount = 0
+        df = pyarrow.orc.read_table(self.filename).to_pandas()
+        for indx, record in df.iterrows():
+            self.assertEqual(
+                "this is a test text",
+                record.get("ftext", None),
+                msg="String data is read correctly",
+            )
+            self.assertEqual(
+                "['test1', 'test2', 'test3', 'test4']",
+                record.get("ftext_array", None),
+                msg="String array data is read correctly",
+            )
+            self.assertEqual(
+                "2.5", record.get("ffloat", None), msg="Float data is read correctly"
+            )
+            self.assertEqual(
+                "10", record.get("fint", None), msg="Int data is read correctly"
+            )
+            self.assertEqual(
+                "False", record.get("fbool", None), msg="Bool data is read correctly"
+            )
+            self.assertEqual(
+                "1582974733.0",
+                record.get("fdatetime", None),
+                msg="DateTime data is read correctly",
+            )
+            rcount += 1
+        self.assertEqual(
+            rcount,
+            10,
+            msg="Number of records read corresponds to number of records written",
+        )
 
 
 if __name__ == "__main__":