+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+* Changed: Removed code comments where the documentation contains the details
 
 ## [0.6.1] - 2025-10-03
 * Added: Provide possibility to specify a location for an Iceberg table (only relevant if a scraper creates the table and it is not already given)